cleanup: remove archived feature wiring and reorganize operation schemas
- Remove tRPC routers for archived features (bestPractices, checklist, codeSnippets, rules)
- Remove archived service factories from app services.ts
- Clean dataProvider.ts: remove all archived resource switch cases
- Archive exportBestPractice.ts (dead code)
- Clean storybookData.ts: remove bestPractice mock data
- Move operation schemas into operation/ subdirectory
