+++ apps/app/src/integrations/trpc/router.ts
@@ -1,9 +1,6 @@
 import { router } from "./init";
 import { agentsRouter } from "./routers/agents";
 import { agentRuntimesRouter } from "./routers/agentRuntimes";
-import { bestPracticesRouter } from "./routers/bestPractices";
-import { checklistRouter } from "./routers/checklist";
-import { codeSnippetsRouter } from "./routers/codeSnippets";
 import { distillationsRouter } from "./routers/distillations";
 import { filesystemRouter } from "./routers/filesystem";
 import { githubProjectsRouter } from "./routers/githubProjects";
@@ -12,7 +9,6 @@ import { operationsRouter } from "./routers/operations";
 import { pipelinesRouter } from "./routers/pipelines";
 import { recipesRouter } from "./routers/recipes";
 import { refinementsRouter } from "./routers/refinements";
-import { rulesRouter } from "./routers/rules";
 import { settingsRouter } from "./routers/settings";
 import { skillsRouter } from "./routers/skills";
 import { operationOutputItemTemplatesRouter } from "./routers/operationOutputItemTemplates";
@@ -25,14 +21,10 @@ export const appRouter = router({
   operations: operationsRouter,
   pipelines: pipelinesRouter,
   settings: settingsRouter,
-  rules: rulesRouter,
-  bestPractices: bestPracticesRouter,
   githubProjects: githubProjectsRouter,
   skills: skillsRouter,
   recipes: recipesRouter,
   refinements: refinementsRouter,
-  checklist: checklistRouter,
-  codeSnippets: codeSnippetsRouter,
   distillations: distillationsRouter,
   operationOutputItemTemplates: operationOutputItemTemplatesRouter,
 });
