+++ apps/app/src/integrations/trpc/services.ts
@@ -2,10 +2,6 @@ import { db } from "@repo/db";
 import {
   createAgentsService,
   createAgentRuntimesService,
-  createBestPracticesBulkService,
-  createBestPracticesService,
-  createChecklistService,
-  createCodeSnippetsService,
   createDistillationsService,
   createGithubProjectsService,
   createJobsService,
@@ -15,7 +11,6 @@ import {
   createPipelinesService,
   createRecipesService,
   createRefinementsService,
-  createRulesService,
   createSettingsService,
   createSkillsService,
   createOperationOutputItemTemplatesService,
@@ -23,10 +18,6 @@ import {
 
 export const agentsService = createAgentsService(db);
 export const agentRuntimesService = createAgentRuntimesService(db);
-export const bestPracticesService = createBestPracticesService(db);
-export const bestPracticesBulkService = createBestPracticesBulkService(db);
-export const checklistService = createChecklistService(db);
-export const codeSnippetsService = createCodeSnippetsService(db);
 export const distillationsService = createDistillationsService(db);
 export const githubProjectsService = createGithubProjectsService(db);
 export const jobsService = createJobsService(db);
@@ -36,7 +27,6 @@ export const pipelinesService = createPipelinesService(db);
 export const pipelineRunnerService = createPipelineRunnerService(db);
 export const recipesService = createRecipesService(db);
 export const refinementsService = createRefinementsService(db);
-export const rulesService = createRulesService(db);
 export const settingsService = createSettingsService(db);
 export const skillsService = createSkillsService(db);
 export const operationOutputItemTemplatesService = createOperationOutputItemTemplatesService(db);
