+++ packages/schemas/src/index.ts
@@ -6,18 +6,13 @@ export * from "./distillation";
 export * from "./ExecutorConfigSchema";
 export * from "./ExecutorTypeSchema";
 export * from "./GithubProjectSchema";
-export * from "./InputPortSchema";
 export * from "./job";
 export * from "./LogLevelSchema";
 export * from "./meta";
-export * from "./OperationConfigSchema";
-export * from "./OperationSchema";
-export * from "./OutputItemSchema";
+export * from "./operation";
 export * from "./RecipeSchema";
 export * from "./refinement";
 export * from "./ScriptLanguageSchema";
 export * from "./SettingsSchema";
 export * from "./SkillSchema";
-export * from "./TemplateContentTypeSchema";
-export * from "./OperationOutputItemTemplateSchema";
 export * from "./pipeline";
