+++ packages/schemas/src/operation/OutputItemSchema.ts
