+++ packages/schemas/src/operation/TemplateContentTypeSchema.ts
